HRMS back/requirements.txt
django>=4.2
djangorestframework>=3.14
djongo>=1.4.0
pymongo>=4.0
django-cors-headers>=3.13
